LibWeb: Add support for generating FlyString parameters from IDL
Previously we would always generate string parameters to pass through
to functions as a `String`. This works work fine if the argument is a
`FlyString const&`, but falls apart for optional types where we need to
accept an `Optional<FlyString> const&`.

Support this by adding support for a [FlyString] extended parameter.
If present, the parameter for the function is generated as a FlyString.