LibJS: Begin using CalendarRecord for AOs
This begins the process of aligning our implementation with the spec
with regard to using CalendarRecord. The main intent here is to make it
much easier to update AOs normative changes.

While this does resolve various FIXMEs, many others above need to be
added in order to be able to pass through a CalendarRecord. The use here
aligns with what I can gather from the spec of what the arguments to
CreateCalendarMethodsRecord should be, but various AOs have been
updated so much it's not completely obvious. Other AOs do not even exist
in the latest version of the spec, but we still rely on them.

Somehow, this commit also fixes two PlainDate roundingmode issues seen
in test262 - tests of which are added in test-js. However, these changes
should make it much easier to fix many more issues in the future :^)

since/calendar-dateuntil-called-with-singular-largestunit.js ❌ -> ✅
until/calendar-dateuntil-called-with-singular-largestunit.js ❌ -> ✅