LibWeb: Factor out a function to generate IDL enumerations
The prototype header generation was a bit long to read for my liking.

This is also a step towards generating code for IDL files only
containing an enum definition without any interface. In that case we
can't put the enum definitions in the prototype header - as there is no
prototype to speak of.